+++ utils/ogmios-client/src/query_ledger_state.rs
@@ -8,7 +8,7 @@ use serde::Deserialize;
 
 pub trait QueryLedgerState: OgmiosClient {
 	#[allow(async_fn_in_trait)]
-	async fn era_summaries(&self) -> Result<Vec<EraSummariesResponse>, OgmiosClientError> {
+	async fn era_summaries(&self) -> Result<Vec<EraSummary>, OgmiosClientError> {
 		self.request("queryLedgerState/eraSummaries", OgmiosParams::empty_positional())
 			.await
 	}
@@ -18,7 +18,7 @@ impl<T: OgmiosClient> QueryLedgerState for T {}
 
 #[derive(Clone, Debug, Deserialize, Eq, PartialEq)]
 #[serde(rename_all = "camelCase")]
-pub struct EraSummariesResponse {
+pub struct EraSummary {
 	pub start: EpochBoundary,
 	pub end: EpochBoundary,
 	pub parameters: EpochParameters,